script: Move 2D canvas stuff into dom/canvas2d folder
