+++ components/script/dom/canvas2d/canvaspattern.rs
