+++ components/script/dom/canvas2d/mod.rs
@@ -1,0 +1,12 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub mod canvas_state;
+pub mod canvasgradient;
+pub mod canvaspattern;
+#[allow(dead_code)]
+pub mod canvasrenderingcontext2d;
+pub mod offscreencanvasrenderingcontext2d;
+pub mod paintrenderingcontext2d;
+pub mod path2d;
