+++ components/script/dom/canvas2d/path2d.rs
