+++ components/script/lib.rs
@@ -32,7 +32,6 @@ pub(crate) mod document_loader;
 #[macro_use]
 mod dom;
 mod canvas_context;
-mod canvas_state;
 pub(crate) mod fetch;
 pub(crate) mod indexed_db;
 mod init;
